### 2023-11-11 #6054 @editorialbot
[PRE REVIEW]: gcamfaostat: An R package to prepare, process, and synthesize FAOSTAT data for global agroeconomic and multisector dynamic modeling
**Submitting author:** @realxinzhao (<a href="http://orcid.org/0000-0002-1801-4393">Xin Zhao</a>)
**Repository:** https://github.com/JGCRI/gcamfaostat
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0
