### 2023-11-13 7f952e6
Add JOSS badge and update refs to include doi as suggested
