+++ README.md
@@ -4,6 +4,7 @@
 [![.github/workflows/draft-pdf.yml](https://github.com/JGCRI/gcamfaostat/actions/workflows/draft-pdf.yml/badge.svg)](https://github.com/JGCRI/gcamfaostat/actions/workflows/draft-pdf.yml)
 [![R](https://github.com/JGCRI/gcamfaostat/actions/workflows/r.yml/badge.svg)](https://github.com/JGCRI/gcamfaostat/actions/workflows/r.yml)
 [![test_coverage](https://github.com/JGCRI/gcamfaostat/actions/workflows/test_coverage.yml/badge.svg)](https://github.com/JGCRI/gcamfaostat/actions/workflows/test_coverage.yml)
+[![status](https://joss.theoj.org/papers/bf4762e8bef702f36b800e0cb4d478a8/status.svg)](https://joss.theoj.org/papers/bf4762e8bef702f36b800e0cb4d478a8)
 <!-- badges: end -->
 
 ### **`gcamfaostat`**: An R package to prepare, process, and synthesize FAOSTAT data for global agroeconomic and multisector dynamic modeling
