+++ R/xfaostat_L201_Forestry.R
@@ -21,8 +21,8 @@ module_xfaostat_L201_Forestry <- function(command, ...) {
     c("FO_Roundwood")
 
   MODULE_OUTPUTS <-
-    c("GCAMDATA_FAOSTAT_ForProdTrade_215Regs_Roundwood_1973to2020",
-      "GCAMDATA_FAOSTAT_ForExportPrice_214Regs_Roundwood_1973to2020")
+    c("For_Balance",
+      "FO_Roundwood_Export_Q_V")
 
   if(command == driver.DECLARE_INPUTS) {
     return(MODULE_INPUTS)
@@ -150,51 +150,21 @@ module_xfaostat_L201_Forestry <- function(command, ...) {
     rm(FO_area, UnitMap, FO_Roundwood, FO_Roundwood_Prod, FO_Roundwood_Import_Q_V)
 
 
-    ## Produce output and export CSV ----
-    GCAMDATA_FAOSTAT_ForProdTrade_215Regs_Roundwood_1973to2020 <-
-      For_Balance %>% filter(element != "Demand") %>%
-      spread(year, value)
-
-    GCAMDATA_FAOSTAT_ForExportPrice_214Regs_Roundwood_1973to2020 <-
-      FO_Roundwood_Export_Q_V %>% spread(year, value)
-
-
-    GCAMDATA_FAOSTAT_ForProdTrade_215Regs_Roundwood_1973to2020 %>%
-      add_title("FAO forestry production, export, and import (roundwood total) by country_year") %>%
+    For_Balance %>%
+      add_title("FAO forestry consumption, production, export, and import (roundwood total) by country_year") %>%
       add_units("m3 ") %>%
       add_comments("Data is preprocessed and generated by gcamdata-FAOSTAT. Gross trade is balanced") %>%
       add_precursors("FO_Roundwood") ->
-      GCAMDATA_FAOSTAT_ForProdTrade_215Regs_Roundwood_1973to2020
+      For_Balance
 
-    GCAMDATA_FAOSTAT_ForExportPrice_214Regs_Roundwood_1973to2020 %>%
+    FO_Roundwood_Export_Q_V %>%
       add_title("FAO forests export qunatity and export value by country_year") %>%
       add_units("m3 and 1000 USD") %>%
       add_comments("Data is generated by gcamdata-FAOSTAT for deriving export prices") %>%
       add_precursors("FO_Roundwood") ->
-      GCAMDATA_FAOSTAT_ForExportPrice_214Regs_Roundwood_1973to2020
+      FO_Roundwood_Export_Q_V
 
 
-    if (OUTPUT_Export_CSV == T) {
-      output_csv_data(
-        "GCAMDATA_FAOSTAT_ForProdTrade_215Regs_Roundwood_1973to2020",
-        col_type_nonyear = "iciccc",
-        title = "FAO forestry production, export, and import (roundwood total) by country_year",
-        unit = "m3",
-        code = "FO",
-        description = "Data is preprocessed and generated by gcamdata-FAOSTAT. Gross trade is balanced",
-        out_dir = DIR_OUTPUT_CSV
-      )
-
-      output_csv_data(
-        "GCAMDATA_FAOSTAT_ForExportPrice_214Regs_Roundwood_1973to2020",
-        col_type_nonyear = "iciccic",
-        title = "FAO forests export qunatity and export value by country_year",
-        unit = "Export Quantity in m3; Export Value in 1000 US$",
-        description = "Data is generated by gcamdata-FAOSTAT for deriving export prices",
-        code = "FO",
-        out_dir = DIR_OUTPUT_CSV
-      )
-    }
 
     return_data(MODULE_OUTPUTS)
 
