+++ R/xfaostat_L101_RawDataPreProc4_FBSH_CB.R
@@ -18,8 +18,8 @@
 module_xfaostat_L101_RawDataPreProc4_FBSH_CB <- function(command, ...) {
 
   MODULE_INPUTS <-
-    c(OPTIONAL_FILE = "aglu/FAO/FAOSTAT/FoodBalanceSheetsHistoric_E_All_Data_(Normalized)_PalceHolder",
-      OPTIONAL_FILE = "aglu/FAO/FAOSTAT/CommodityBalances_(non-food)_E_All_Data_(Normalized)_PalceHolder",
+    c(FAOSTAT_FILE = "aglu/FAO/FAOSTAT/FoodBalanceSheetsHistoric_E_All_Data_Normalized",
+      FAOSTAT_FILE = "aglu/FAO/FAOSTAT/CommodityBalances_(non-food)_E_All_Data_Normalized",
       "QCL_area_code_map")
 
   MODULE_OUTPUTS <-
@@ -181,8 +181,8 @@ module_xfaostat_L101_RawDataPreProc4_FBSH_CB <- function(command, ...) {
       add_title("FAO FBSH and CB, food and commodity balance before 2013, wide", overwrite = T) %>%
       add_units("1000 tonne") %>%
       add_comments("Preprocessed FAO FBSH_CB") %>%
-      add_precursors("aglu/FAO/FAOSTAT/FoodBalanceSheetsHistoric_E_All_Data_(Normalized)_PalceHolder",
-                     "aglu/FAO/FAOSTAT/CommodityBalances_(non-food)_E_All_Data_(Normalized)_PalceHolder",
+      add_precursors("aglu/FAO/FAOSTAT/FoodBalanceSheetsHistoric_E_All_Data_Normalized",
+                     "aglu/FAO/FAOSTAT/CommodityBalances_(non-food)_E_All_Data_Normalized",
                      "QCL_area_code_map") ->
       FBSH_CB_wide
 
