### 2023-11-20 88b33ac
rm old data inputs for AgLU modules; new data will be added later
