+++ R/xfaostat_L199_CSVExportAgSUA.R
@@ -35,7 +35,20 @@ module_xfaostat_L199_CSVExportAgSUA <- function(command, ...) {
       "RFN_ProdDemand")
 
   MODULE_OUTPUTS <-
-    c("xfaostat_L199_GCAMFAOSTAT_CSV")
+    c("xfaostat_L199_GCAMFAOSTAT_CSV",
+      CSV = "GCAMFAOSTAT_SUA",
+      CSV = "GCAMFAOSTAT_BiTrade",
+      CSV = "GCAMFAOSTAT_FBSH_CB",
+      CSV = "GCAMFAOSTAT_NonFodderProdArea",
+      CSV = "GCAMFAOSTAT_FodderProdArea",
+      CSV = "GCAMFAOSTAT_AnimalStock",
+      CSV = "GCAMFAOSTAT_ProdPrice",
+      CSV = "FAO_GDP_Deflators",
+      CSV = "GCAMFAOSTAT_MacroNutrientRate",
+      CSV = "GCAMFAOSTAT_ForProdTrade",
+      CSV = "GCAMFAOSTAT_ForExpPrice",
+      CSV = "GCAMFAOSTAT_LandCover",
+      CSV = "GCAMFAOSTAT_NFertilizer")
 
   if(command == driver.DECLARE_INPUTS) {
     return(MODULE_INPUTS)
@@ -52,30 +65,28 @@ module_xfaostat_L199_CSVExportAgSUA <- function(command, ...) {
     all_data <- list(...)[[1]]
 
 
+    Curr_Envir <- environment()
+
     # adding dummy output ----
     xfaostat_L199_GCAMFAOSTAT_CSV <-
       tibble(CSV_export =
-               c("GCAMFAOSTAT_SUA",
-                 "GCAMFAOSTAT_BiTrade",
-                 "GCAMFAOSTAT_FBSH_CB",
-                 "GCAMFAOSTAT_NonFodderProdArea",
-                 "GCAMFAOSTAT_FodderProdArea",
-                 "GCAMFAOSTAT_AnimalStock",
-                 "GCAMFAOSTAT_ProdPrice",
-                 "FAO_GDP_Deflators",
-                 "GCAMFAOSTAT_MacroNutrientRate",
-                 "GCAMFAOSTAT_ForProdTrade",
-                 "GCAMFAOSTAT_ForExpPrice",
-                 "GCAMFAOSTAT_LandCover",
-                 "GCAMFAOSTAT_NFertilizer"))
+               MODULE_OUTPUTS)
 
 
-    if (OUTPUT_Export_CSV == T) {
+    if (OUTPUT_Export_CSV == FALSE) {
+
+    lapply(MODULE_OUTPUTS[MODULE_OUTPUTS %>% names() == "CSV"],
+           function(output){
+             assign(output, empty_data() %>%
+                      add_title(output) %>%
+                      add_precursors("xfaostat_L199_GCAMFAOSTAT_CSV"), envir =  Curr_Envir)
+           })
+    }
+    else {
 
       # Load required inputs ----
       get_data_list(all_data, MODULE_INPUTS, strip_attributes = TRUE)
 
-      Curr_Envir <- environment()
 
       # Helper functions
       xfaostat_L199_GCAMFAOSTAT_CSV <- NULL
@@ -128,13 +139,18 @@ module_xfaostat_L199_CSVExportAgSUA <- function(command, ...) {
           SUDAN2012_MERGE = T
         ) %>%
         filter(value != 0.0) %>%
-        transmute(area_code, item_code, source_code, year, value) ->
-        DF_TEMP
+        transmute(area_code, item_code, source_code, year, value) %>%
[…]
