+++ R/xfaostat_L105_DataConnectionToSUA.R
@@ -19,7 +19,6 @@ module_xfaostat_L105_DataConnectionToSUA <- function(command, ...) {
 
   MODULE_INPUTS <-
     c(FILE = "aglu/FAO/FAO_items",
-      FILE = "aglu/FAO/Mapping_FBSH_SCL_OilCake",
       "QCL_PROD",
       "QCL_AN_LIVEANIMAL_MEATEQ",
       "TCL_wide",
@@ -40,7 +39,7 @@ module_xfaostat_L105_DataConnectionToSUA <- function(command, ...) {
     year <- value <- Year <- Value <- FAO_country <- iso <- NULL    # silence package check.
     SCL_wide <- element_code <- element <- area_code <- item_code <- area <-
       item <- unit <- FBS_wide <- FBSH_CB_wide <- TCL_wide <- TM_bilateral_wide <-
-      QCL_PROD <- FAO_items <- tier <- QCL <- Mapping_FBSH_SCL_OilCake <- oil <-
+      QCL_PROD <- FAO_items <- tier <- QCL <- oil <-
       cake <- SCL_item_oil <- SCL_item_cake <- cake_rate <- cake_rate_world <-
       DS_key_coproduct_item <- Production <- Import <- Export <- DS_demand <-
       DS_production <- CoproductRate <- QCL_AN_LIVEANIMAL_MEATEQ <- `Closing stocks` <-
@@ -143,8 +142,11 @@ module_xfaostat_L105_DataConnectionToSUA <- function(command, ...) {
              element = replace(element, element == "Losses", "Loss"),
              element = replace(element, element == "Processing", "Processed")) %>%
       # convert units back to tonnes first since FBS originally used 1000 tons
-      mutate(value = value * 1000, unit = "tonnes")->
+      mutate(value = value * 1000, unit = "tonnes") ->
       FBS
+
+    # FBS has all needed elements in FBS_element_new
+    assertthat::assert_that(setdiff(SCL_element_new, unique(FBS$element)) == "Opening stocks")
 
     ## 1.6. Define balance elements ----
 
@@ -168,7 +170,6 @@ module_xfaostat_L105_DataConnectionToSUA <- function(command, ...) {
     # Update area code in QCL
     QCL_PROD %>% filter(year %in% FAOSTAT_Hist_Year_FBS) %>%  distinct(area_code) %>% pull ->
       QCL_area_code_FBS
-
 
     # 2. Create helper functions to simplify join by data set ----
 
@@ -467,99 +468,29 @@ module_xfaostat_L105_DataConnectionToSUA <- function(command, ...) {
 
     assert_FBS_balance(.DF = Bal_new_tier4)
 
+
     ## 3.5 Bal_new_tier5 ----
-    # Tier5 includes 10 oilseed cake items that were omitted in FAOSTAT
-    # Note  key oilseed cake data are not currently provided by FAOSTAT after 2013.
-    # We extrapolate the data based on oil-cake rates in 2010 to 2013
+    #Tier7 includes 12 fish items from FBS and FBSH. Item code came from FBS as well
 
-    ### 3.5.1 Get oil cake production data based on cake rate from CB_FBS_CakeRate ----
-
-    # Get the mapping of oil and cake items between FBSH, CB and SCL
-    # read in already
-    #Mapping_FBSH_SCL_OilCake <- readr::read_csv(file.path("inst/extdata/aglu/FAO", "Mapping_FBSH_SCL_OilCake.csv"), comment = "#")
-
-    # The reference period is 2011: 2013 for cake rate calculation
-    # Merge oil and cake data
-    Mapping_FBSH_SCL_OilCake %>%
[…]
