+++ R/xfaostat_L999_CSVExport.R
@@ -1,0 +1,557 @@
+# Copyright 2019 Battelle Memorial Institute; see the LICENSE file.
+
+#' module_xfaostat_L999_CSVExport
+#'
+#' Generate supply utilization balance in primary equivalent
+#'
+#' @param command API command to execute
+#' @param ... other optional parameters, depending on command
+#' @return Depends on \code{command}: either a vector of required inputs, a vector of output names, or (if
+#'   \code{command} is "MAKE") all the generated outputs
+#' @details This chunk compiles balanced supply utilization data in primary equivalent in GCAM region and commodities.
+#' @importFrom assertthat assert_that
+#' @importFrom dplyr summarize bind_rows filter if_else inner_join left_join mutate rename select n group_by_at
+#' first case_when vars
+#' @importFrom tibble tibble
+#' @importFrom tidyr complete drop_na gather nesting spread replace_na
+#' @author XZ Nov2023
+module_xfaostat_L999_CSVExport <- function(command, ...) {
+
+  MODULE_INPUTS <-
+    c("PD",
+      "Bal_new_all",
+      "FBSH_CB_wide",
+      "QCL_PROD",
+      "QCL_AN_LIVEANIMAL",
+      "QCL_AN_PRIMARY_MILK",
+      "QCL_CROP_PRIMARY",
+      "QCL_FODDERCROP",
+      "QCL_PRIMARY_PROD_PV",
+      "TM_bilateral_wide",
+      "SUA_food_macronutrient_rate",
+      "For_Balance",
+      "FO_RoundwoodProducts_Export_Q_V",
+      "RL_LandCover",
+      "RFN_ProdDemand")
+
+  MODULE_OUTPUTS <-
+    c("xfaostat_L199_GCAMFAOSTAT_CSV",
+      CSV = "GCAMFAOSTAT_SUA",
+      CSV = "GCAMFAOSTAT_BiTrade",
+      CSV = "GCAMFAOSTAT_FBSH_CB",
+      CSV = "GCAMFAOSTAT_NonFodderProdArea",
+      CSV = "GCAMFAOSTAT_FodderProdArea",
+      CSV = "GCAMFAOSTAT_AnimalStock",
+      CSV = "GCAMFAOSTAT_ProdPrice",
+      CSV = "FAO_GDP_Deflators",
+      CSV = "GCAMFAOSTAT_MacroNutrientRate",
+      CSV = "GCAMFAOSTAT_ForProdTrade",
+      CSV = "GCAMFAOSTAT_ForExpPrice",
+      CSV = "GCAMFAOSTAT_LandCover",
+      CSV = "GCAMFAOSTAT_NFertilizer")
+
+  if(command == driver.DECLARE_INPUTS) {
+    return(MODULE_INPUTS)
+  } else if(command == driver.DECLARE_OUTPUTS) {
+    return(MODULE_OUTPUTS)
+  } else if(command == driver.MAKE) {
+
+    year <- value <- Year <- Value <- FAO_country <- iso <- NULL    # silence package check.
+    area_code <- source_code <- area_code1 <- item_code <- Bal_new_all <- element <- FBSH_CB_wide <-
+      element_code <- QCL_CROP_PRIMARY <- QCL_PROD <- unit <- QCL_FODDERCROP <- QCL_AN_LIVEANIMAL <-
+      QCL_AN_PRIMARY_MILK <- QCL_PRIMARY_PROD_PV <- area <- PD <- SUA_food_macronutrient_rate  <-
+      TM_bilateral_wide <- NULL
+
+    all_data <- list(...)[[1]]
+
+
+    Curr_Envir <- environment()
+
+    # adding dummy output ----
[…]
