### 2023-11-27 #6054 @kthyng
@openjournals/dev Why isn't the list of similar papers coming up for this submission?
