### 2023-12-09 #6054 @arfon
@kthyng – this is a bug with how we generate the papers on GitHub. As an interim solution, I've generated this list locally:

The top 5 most similar papers are:
[geostan: An R package for Bayesian spatial analysis](https://github.com/openjournals/joss-reviews/issues/4716)
Submitting author: `@ConnorDonegan`
Handling editor: `@jbytecode (Active)`
Reviewers: `@wcjochem, @becarioprecario`
Similarity score: 0.8179978544494104

[grapesAgri1: Collection of Shiny Apps for Data Analysis in Agriculture](https://github.com/openjournals/joss-reviews/issues/3437)
Submitting author: `@pratheesh3780`
Handling editor: `@fabian-s (Active)`
Reviewers: `@adithirgis, @elimillera`
Similarity score: 0.8132841262870389

[StatAid: An R package with a graphical user interface for data analysis](https://github.com/openjournals/joss-reviews/issues/2630)
Submitting author: `@VincentAlcazer`
Handling editor: `@mikldk (Retired)`
Reviewers: `@nistara, @adithirgis`
Similarity score: 0.8099523747873361

[rnassqs: An R package to access agricultural data via the USDA National Agricultural Statistics Service (USDA-NASS) 'Quick Stats' API](https://github.com/openjournals/joss-reviews/issues/1880)
Submitting author: `@potterzot`
Handling editor: `@kyleniemeyer (Active)`
Reviewers: `@kyleniemeyer`
Similarity score: 0.8092332754201275

[Visualizations with statistical details: The 'ggstatsplot' approach](https://github.com/openjournals/joss-reviews/issues/3167)
Submitting author: `@IndrajeetPatil`
Handling editor: `@csoneson (Active)`
Reviewers: `@njtierney, @kevinrue`
Similarity score: 0.8060231737780731
