+++ pkgdown.yml
@@ -8,7 +8,7 @@ articles:
   vignette_processing_flow: vignette_processing_flow.html
   vignette_use_cases: vignette_use_cases.html
   vignette_visualization: vignette_visualization.html
-last_built: 2023-11-13T04:59Z
+last_built: 2024-01-28T14:16Z
 urls:
   reference: https://jgcri.github.io/gcamfaostat/reference
   article: https://jgcri.github.io/gcamfaostat/articles
