### 2024-02-01 #6054 @realxinzhao
Dear @hugoledoux,

Thank you for editing this paper.  Here is a list of potential reviewers:

1. tscheypidi (via Reviewers Management System; no conflict of interest)
2. KristinaRiemer (via Reviewers Management System; no conflict of interest)
3. Claudia Rodes: [claudia.rodes@bc3research.org](mailto:claudia.rodes@bc3research.org) (via research network; no conflict of interest)
4. [jiangyongye](https://github.com/jiangyongye) (gcam research network; no conflict of interest)

Please let us know if we can provide more information to facilitate this process.
