### 2024-02-02 86d50df
Adding figures for food processing paper
