+++ vignettes/vignette_visualization.Rmd
@@ -461,6 +461,232 @@ See [Cast Study](https://jgcri.github.io/gcamfaostat/articles/vignette_use_cases
 ### 7. Primary equivalent [in process]
 
 
+### 8. Primary ag to final uses
+* Get data ready
+```{r, eval =FALSE}
+library(ggplot2)
+library(dplyr)
+library(patchwork)
+library(RColorBrewer)
+
+devtools::load_all()
+"L100.FAO_SUA_APE_balance" %>% load_from_cache() %>% first -> L100.FAO_SUA_APE_balance
+"L100.FAO_PRODSTAT_TO_DOWNSCAL" %>% load_from_cache() %>% first -> L100.FAO_PRODSTAT_TO_DOWNSCAL
+"FAO_Food_Macronutrient_All_2010_2019" %>% load_from_cache() %>% first -> FAO_Food_Macronutrient_All_2010_2019
+
+c("Opening stocks", "Import", "Production") -> ElEMSupply
+c("Food", "Feed",  "Processed", "Other uses", "Export", "Closing stocks", "Residuals") -> ELEMDemand
+c(ElEMSupply, ELEMDemand) -> ELEMAll
+
+# Define color pal.
+c(brewer.pal(n = length(ELEMAll), name = "Spectral")[1:length(ElEMSupply)],
+  brewer.pal(n = length(ELEMAll), name = "Spectral")[length(ElEMSupply)+1:length(ELEMDemand)]
+) -> ColUpdate
+
+GCAMCrops <-
+  c( "Corn", "Wheat", "Rice", "OtherGrain", "RootTuber",  "Soybean" ,
+    "OilCrop",        "OilPalm",  "SugarCrop",      "Vegetables",  "Fruits",
+    "NutsSeeds", "Legumes", "FiberCrop", "MiscCrop")        
+GCAMLivestock <-
+  c("Beef", "Dairy", "Pork", "Poultry", "SheepGoat", "OtherMeat_Fish")
+
+L100.FAO_SUA_APE_balance %>% 
+  filter(!GCAM_commodity %in% GCAMLivestock) %>% 
+  filter(year == 2015) %>% 
+  mutate(region = if_else(GCAM_region_ID == 1, "USA", "ROW")) %>% 
+  group_by(GCAM_commodity, region, element) %>% 
+  summarize(value = sum(value)) %>% ungroup() %>% 
+  filter(element != "Net_Export") %>% 
+  mutate(value = if_else(element %in% c("Export", "Other uses", "Food", "Feed"), -value, value)) %>% 
+  mutate(region = factor(region, levels = c("USA", "ROW"))) %>% 
+  mutate(GCAM_commodity = factor(GCAM_commodity, levels = GCAMCrops)) %>% 
+  mutate(element = factor(element, levels = c("Import", "Production", "Export", "Other uses", "Feed", "Food"))) %>% 
+  ggplot() + facet_wrap(~region, scales = "free") +
+  geom_bar(aes(x = GCAM_commodity, y = value, fill = element),
+           stat = "identity", color = "black", position="stack") +
+  geom_hline(yintercept = 0) +
+  scale_fill_brewer(palette = "Spectral") +
+  labs(x = "GCAM Crop Commodities",
+       y = "Million tonnes", fill = "Element",
+       title = "(A) Supply utilization accounts in primary equivalent for GCAM crop commodities",
+       #subtitle = "Note: other uses include seed use, losses, storage variation, and industrial use",
+       #caption = "Data: FAOSTAT \n Package: JGCRI/gcamfaostat"
+       ) +
+  theme_bw() + 
+  theme(text = element_text(size = 16),
+        axis.text = element_text(colour = "black", size = 16),
+        axis.text.x = element_text(angle = 35, hjust = 1, vjust = 1),
[…]
