### 2024-02-11 bd77876
Update figures for the trade discussion in use case
