### 2024-02-28 JGCRI/gcamfaostat#7 @klau506
Manuscript Suggestions
https://github.com/openjournals/joss-reviews/issues/6388

- [x] L 51: "for GCAM" or "for the GCAM model"
- [x] L 80: "from FAOSTAT"
- [x] L 171: "it inherits"
- [x] L 218: maybe detail what `area_code` is ("FAO regions using the area_code ID"?) or how to get it. From which mapping file?
- [x] L 219: "dissolutions"
- [x] L 244: add space before "Throught"
- [x] Captions: some captions start with "The" (e.g. Figure 1, Figure2, Figure 3), and some others with directly the description (e.g. Figure 4, Table 1). Might be better to homogenize
- [x] Consider inline code formatting when referring to file names and modules, such as `module_xfaostat_L105_DataConnectionToSUA` or `constants.R`. In some cases it is already used, but in some others this notation is missing
- [x] In the *Statement of need* section, you mention the *transparency and reproducibility*, which I think are key elements of this package. It would be nice to find some reference, mention, or linkage to these concepts throughout the text.
