### 2024-02-28 JGCRI/gcamfaostat#8 @klau506
Documentation Suggestions
https://github.com/openjournals/joss-reviews/issues/6388

- [ ] In the section [Getting Started](https://jgcri.github.io/gcamfaostat/articles/vignette_getting_started.html#r), in the first step of *clone this repository*, put the command `git clone ...` in a new line, so that the user can copy it directly and it keeps the same format as the commands detailed in the *loading the gcamfaostat pkg* or *pkg dependencies* subsections
- [ ] [Quick start in R](https://github.com/jgcri/gcamfaostat/?tab=readme-ov-file#quick-start-in-r--40--rstudio) does not match the [Getting Started](https://jgcri.github.io/gcamfaostat/articles/vignette_getting_started.html#r) section. In one it is specified that you should clone the repo, in the other it is not stated. Just to make sure that both options work fine. In the same line, the second point of the [Quick start in R](https://github.com/jgcri/gcamfaostat/?tab=readme-ov-file#quick-start-in-r--40--rstudio) section is somehow confusing when it asks you to open the `gcamfaostat.Rproj`, but does not tell you how to get it
- [ ] In the section [Getting Started](https://jgcri.github.io/gcamfaostat/articles/vignette_getting_started.html#r), when cloning the repo, I was in doubt where to clone it: inside the GCAM model folder, like `gcamdata`, or outside? Could be helpful to detail where it is better to clone it (maybe depending on your user case?) or if there is only one option, detail. (FYI: I decided to clone it outside the `GCAM` folder)
- [ ] When doing the following steps, I ran into an error:
```
> devtools::load_all()
> devtools::install_deps() # only XML (3.99-0.14 -> 3.99-0.16.1) [CRAN] appears, but I'm having some problems to update it in my laptop
> driver_drake()
Loading required namespace: drake
GCAM Data System v1.0.0
Found 19 chunks
Found 77 chunk data requirements
Found 43 chunk data products
37 chunk data input(s) not accounted for
Error in `mutate()`:
i In argument: `command = paste0(PACKAGE_NAME, "::PREBUILT_DATA[['", target, "']]")`.
Caused by error:
! `command` must be size 0 or 1, not 99.
Run `rlang::last_trace()` to see where the error occurred.
```
I think this is due to the `dplyr` or `tidyr` pkg version. Maybe it would be good to specify the required pkg version, if not with the `devtools::install_deps()`, at least in a sentence to avoid confusion.
- [ ] Suggestion: In the [Prepare data](https://jgcri.github.io/gcamfaostat/articles/vignette_preparing_data.html) section, include examples of how to run the functions described in sections 3-5 to help the user to understand how to find/specify de DATASETCODE
- [ ] In the section [User cases FAOSTAT raw data](https://jgcri.github.io/gcamfaostat/articles/vignette_use_cases.html#faostat-raw-data-and-processing-output-updates), it would be helpful to detail how to change the *year variables*, since there are many variables that are year related. Maybe include an example? Also, for users not familiarised with the `gcamdata` pkg, it might be helpful to specify the path where `constants.R` is located
- [ ] In the [User cases Country aggregation](https://jgcri.github.io/gcamfaostat/articles/vignette_use_cases.html#country-aggregation-and-disaggregation), would be helpful to detail how to obtain the `area_code` in the `FAOSTAT_AREA_RM_NONEXIST` function
- [ ] In the [bilateral trade case study](https://jgcri.github.io/gcamfaostat/articles/vignette_use_cases.html#case-study-wheat-bilateral-trade-pattern-changes), when considering the download of the latest TM, you could include the command
```
devtools::load_all()

# Download the latest TM (bilateral trade) from FAOSTAT
# To the path specified in DIR_RAW_DATA_FAOSTAT (see constants.R)
# HERE?
...
```
- [ ] In the [bilateral trade case study](https://jgcri.github.io/gcamfaostat/articles/vignette_use_cases.html#case-study-wheat-bilateral-trade-pattern-changes)
I found this message
```
  > FF_download_FAOSTAT(DATASETCODE = "TM",
+                     DATA_FOLDER = file.path("..",DIR_RAW_DATA_FAOSTAT),
+                     OverWrite = TRUE)
Starting download dataTM ...
Error in utils::download.file(url = DatasetURL, destfile = zip_file_name) :
  cannot open destfile '../inst/extdata/aglu/FAO/FAOSTAT/Trade_DetailedTradeMatrix_E_All_Data_(Normalized).zip', reason 'No such file or directory'
> getwd()
[1] "C:/gcamfaostat"
```
and I made sure that in `../inst/extdata/aglu/FAO/FAOSTAT/Trade_DetailedTradeMatrix_E_All_Data_(Normalized).zip` I have the file. The only element that might be different, is that it has the *WinRAR ZIP archive (.zip)* extension
