### 2024-02-28 JGCRI/gcamfaostat#9 @klau506
Code Suggestions
https://github.com/openjournals/joss-reviews/issues/6388

- [x] When I downloaded the raw data from FAOSTAT according to [this section](https://jgcri.github.io/gcamfaostat/articles/vignette_preparing_data.html#download-faostat-raw-data), a warning appeared
```
> FF_download_RemoteArchive(gcamfaostat_metadata(OnlyReturnDatasetCodeRequired = T))
Starting download dataQCL ...
* Starting download dataPP ...
* Starting download dataPD ...
* Starting download dataTCL ...
* Starting download dataTM ...
* Starting download dataSCL ...
* Starting download dataFBS ...
* Starting download dataFBSH ...
* Starting download dataCB ...
* Starting download dataRFN ...
* Starting download dataRL ...
* Starting download dataFO ...
* Starting download dataOA ...
trying URL 'https://zenodo.org/record/8260225/files/SUA_Crops_Livestock_E_All_Data_(Normalized).zip'
Content type 'application/octet-stream' length 64169556 bytes (61.2 MB)
downloaded 61.2 MB

Download complete.
Warning message:
There was 1 warning in `filter()`.
i In argument: `datasetcode == DATASETCODE`.
Caused by warning in `datasetcode == DATASETCODE`:
! longer object length is not a multiple of shorter object length
```
and only `SUA_Crops_Livestock_E_All_Data_(Normalized)` was downloaded. It seems that the function does not accept a vector as `DATASETCODE`. Either fix it or document accordingly
- [x] Some warnings and notes when performing `devtools::check()` could be easily removed, such as
```
W  checking dependencies in R code (1.5s)
   Unexported object imported by a ':::' call: 'utils:::format.object_size'
     See the note in ?`:::` about the use of this operator.
     Including base/recommended package(s):
     'utils'

N  checking R code for possible problems (13.4s)
   SUA_bal_adjust: no visible binding for global variable 'Tourist
     consumption'
   driver_drake: no visible binding for '<<-' assignment to
     'xml.XML_SUFFIX'
     Tourist consumption
```
- [x] #12
