### 2024-02-28 #6388 @klau506
Dear @hugoledoux and @realxinzhao,

This is an interesting contribution with the potential to be a broadly useful resource and source of documentation. I have read the manuscript and the documentation, installed the software and examined its code; see specific suggestions about both referenced above. Broadly:

- The manuscript is generally well written but it could benefit from minor modifications. Further details [here](https://github.com/JGCRI/gcamfaostat/issues/7)
- The documentation is extensive but there are some aspects that may need further precision or clarification. Futher details [here](https://github.com/JGCRI/gcamfaostat/issues/8)
- The code is well structured and organized. Running `devtools::check()` on the `Rproj` currently generates some notes and warnings, some of which could be easily removed. Further details [here](https://github.com/JGCRI/gcamfaostat/issues/9)

I hope this is useful! I appreciate that the authors have done a lot of work, and am happy to clarify/answer questions if useful.

Best,
Clàudia
