### 2024-03-05 JGCRI/gcamfaostat#10 @HenriKajasilta
Usability Fixes
Followed the code junks in the [process flow](https://jgcri.github.io/gcamfaostat/articles/vignette_processing_flow.html#examples) and [visualisation](https://jgcri.github.io/gcamfaostat/articles/vignette_visualization.html#code-for-generating-figures) sections: in both, the bottleneck seemed to be the `load_from_cache()` function, which produced an error. If I understand correctly, the primary problem is with the `driver_drake()` function, as it produced the error mentioned in this https://github.com/JGCRI/gcamfaostat/issues/8 issue. Since `driver_drake()` is not working as expected, there is also problem what is stored in cache. Calling a `driver()` function seems to work and downloads a bunch of different files, but doesn't seem to produce a data trace.
```
> MODULE_INPUTS %>% load_from_cache() -> all_data
Error in (function (target, character_only = FALSE, path = NULL, search = NULL,  :
  cannot find drake cache.
```

- [x] Fix the aforementioned workflow.

I would prefer examples in functions. Especially since the [driver()](https://jgcri.github.io/gcamdata/reference/driver.html ) function has many options, it would be beneficial to have an option set, that is easy and fast to run, so that the user gets a better illustration of what data has been processed.

- [x] Suggestion: Create an example of how to use the function(s), that is/are easy to perform and light to run.

Changing the output directory to where the csv files are written (as mentioned [here](https://jgcri.github.io/gcamfaostat/articles/vignette_getting_started.html#output-files), will produce some of the outputs in the specified location, but will also produce the same files under the `./output`. I assume this is not intended.

- [x] Fix data storing to multiple locations.

There are packages, which are not mentioned in [DESCRIPTIONS](https://github.com/JGCRI/gcamfaostat/blob/main/DESCRIPTION). They may not be necessary for the package to work in some parts, but when working for example with the visualizations more packages were necessary to install. Take care that at least `devtools::check()` will not produce any errors and warnings.

- [x] Fix the issue so that building the package won't produce an error.
