### 2024-03-05 JGCRI/gcamfaostat#11 @HenriKajasilta
Comments on the paper
The functionality and intended use of the package are clearly stated. However, the length of the paper is well over 1000 words as instructed. The chapter on _Design and Functionality_ could probably be summarised with a high-level illustration of _preparation_, _processing_ and _visualisation_. The detailed overview of these stages can be found on the gh-pages anyway.

- [x] Check and fix the length of the paper.

It wasn't really mentioned if there are similar applications/packages available. Based on the paper, at least not as comprehensive as `gcamfaostat`, but it would have been interesting to have some references, if something similar is available. However, the paper states: "_It is noteworthy that each “agroeconomic modeling team typically develops its own assumptions and methods to prepare and process FAOSTAT data_", which I interpret to mean that most of the data processing is not done by off-the-shelf systems.

- [x] Suggestion: Indicate if there are other packages which are similar to `gcamfaostat`.
