### 2024-03-05 #6388 @HenriKajasilta
Hi @hugoledoux and @realxinzhao,

I have installed the `gcamfaostat` package and familiarised myself with the documentation and manuscript. The package clearly has a purpose in its respective field.

I have proposed fixes and suggestions that would improve the usability of the package and help it to meet the JOSS criteria. The resulting issues are here: [#10](https://github.com/JGCRI/gcamfaostat/issues/10)  and [#11](https://github.com/JGCRI/gcamfaostat/issues/11)

Please, let me know if you want me to clarify some of the comments or if I have misunderstood something in the package. I will be happy to take another look at the functionality once the fixes have been made!

Br,
Henri
