+++ vignettes/references.bib
@@ -345,3 +345,44 @@ publisher: Nature Publishing Group},
   publisher={Nature Publishing Group},
   DOI={10.1038/sdata.2016.18}
 }
+@article{zhaoLandCDR2024,
+	title = {Trade-offs in land-based carbon removal measures under 1.5 °{C} and 2 °{C} futures},
+	volume = {15},
+	copyright = {2024 The Author(s)},
+	issn = {2041-1723},
+	url = {https://www.nature.com/articles/s41467-024-46575-3},
+	doi = {10.1038/s41467-024-46575-3},
+	abstract = {Land-based carbon removals, specifically afforestation/reforestation and bioenergy with carbon capture and storage (BECCS), vary widely in 1.5 °C and 2 °C scenarios generated by integrated assessment models. Because underlying drivers are difficult to assess, we use a well-known integrated assessment model, GCAM, to demonstrate that land-based carbon removals are sensitive to the strength and scope of land-based mitigation policies. We find that while cumulative afforestation/reforestation and BECCS deployment are inversely related, they are both typically part of cost-effective mitigation pathways, with forestry options deployed earlier. While the CO2 removal intensity (removal per unit land) of BECCS is typically higher than afforestation/reforestation over long time horizons, the BECCS removal intensity is sensitive to feedstock and technology choices whereas the afforestation/reforestation removal intensity is sensitive to land policy choices. Finally, we find a generally positive relationship between agricultural prices and removal effectiveness of land-based mitigation, suggesting that some trade-offs may be difficult to avoid.},
+	language = {en},
+	number = {1},
+	urldate = {2024-03-14},
+	journal = {Nature Communications},
+	author = {Zhao, Xin and Mignone, Bryan K. and Wise, Marshall A. and McJeon, Haewon C.},
+	month = mar,
+	year = {2024},
+	note = {Publisher: Nature Publishing Group},
+	keywords = {Climate-change mitigation, Climate-change policy, Environmental impact, Geography},
+	pages = {2297},
+	file = {Full Text PDF:C\:\\Users\\zhao752\\Zotero\\storage\\I9GGJKZ8\\Zhao et al. - 2024 - Trade-offs in land-based carbon removal measures u.pdf:application/pdf},
+}
+
+@article{wilkinson_fair_2016,
+	title = {The {FAIR} {Guiding} {Principles} for scientific data management and stewardship},
+	volume = {3},
+	copyright = {2016 The Author(s)},
+	issn = {2052-4463},
+	url = {https://www.nature.com/articles/sdata201618},
+	doi = {10.1038/sdata.2016.18},
+	abstract = {There is an urgent need to improve the infrastructure supporting the reuse of scholarly data. A diverse set of stakeholders—representing academia, industry, funding agencies, and scholarly publishers—have come together to design and jointly endorse a concise and measureable set of principles that we refer to as the FAIR Data Principles. The intent is that these may act as a guideline for those wishing to enhance the reusability of their data holdings. Distinct from peer initiatives that focus on the human scholar, the FAIR Principles put specific emphasis on enhancing the ability of machines to automatically find and use the data, in addition to supporting its reuse by individuals. This Comment is the first formal publication of the FAIR Principles, and includes the rationale behind them, and some exemplar implementations in the community.},
+	language = {en},
+	number = {1},
+	urldate = {2024-03-15},
+	journal = {Scientific Data},
[…]
