### 2024-03-15 realxinzhao/gcamfaostat#7 @realxinzhao
Thank you so much, @klau506, for the careful reading and great suggestions.
All the suggestions are accepted or addressed accordingly.
For the last comment on `transparency and reproducibility`, we added a reference to the seminal work by Wilkinson et al. ([2016](doi.org/10.1038/sdata.2016.18)). We agree this is important, though there are not many studies emphasizing the importance of transparency and reproducibility for global economic & multisector modeling.
We will address the remaining comments and keep you posted.
