### 2024-03-15 realxinzhao/gcamfaostat#8 @realxinzhao
@klau506 Thanks again for the super valuable comments and careful reading of the documentation! All suggestions are accepted. Point-by-point responses are also provided below. Please let us know if we can provide additional clarifications. Thank you!

> In the section [Getting Started](https://jgcri.github.io/gcamfaostat/articles/vignette_getting_started.html#r), in the first step of clone this repository, put the command git clone ... in a new line, so that the user can copy it directly and it keeps the same format as the commands detailed in the loading the gcamfaostat pkg or pkg dependencies subsections

Done. Thanks!

>  [Quick start in R](https://github.com/jgcri/gcamfaostat/?tab=readme-ov-file#quick-start-in-r--40--rstudio) does not match the [Getting Started](https://jgcri.github.io/gcamfaostat/articles/vignette_getting_started.html#r) section. In one it is specified that you should clone the repo, in the other it is not stated. Just to make sure that both options work fine. In the same line, the second point of the [Quick start in R](https://github.com/jgcri/gcamfaostat/?tab=readme-ov-file#quick-start-in-r--40--rstudio) section is somehow confusing when it asks you to open the gcamfaostat.Rproj, but does not tell you how to get it

Good points! They were revised/clarified in both places. `Git clone` is usually preferred as it specifies the location where the package is downloaded.

> Maybe it would be good to specify the required pkg version, if not with the devtools::install_deps(), at least in a sentence to avoid confusion.

Good catch!  I think the issue came from either tibble or dplyr package.
And I was wrong previously about the issue as it came up again.
I realize `devtools::install_deps()` only updates packages per the specifications in the DESCRIPTION file. And we only required lower versions of the packages mostly following the original settings in gcamdata. We have now updated the package versions in DESCRIPTION.
We also keep the `renv` approach since it also recorded package versions. This sentence is added: "`renv` has been initialized and a `renv.lock` file was included."

> In the section [Getting Started](https://jgcri.github.io/gcamfaostat/articles/vignette_getting_started.html#r), when cloning the repo, I was in doubt where to clone it: inside the GCAM model folder, like gcamdata, or outside? Could be helpful to detail where it is better to clone it (maybe depending on your user case?) or if there is only one option, detail. (FYI: I decided to clone it outside the GCAM folder)

Thanks. This sentence is added to clarify this: "Note that `gcamfaostat` is not dependent on GCAM, so it can be installed to folders as desired by the user."

> Suggestion: In the [Prepare data](https://jgcri.github.io/gcamfaostat/articles/vignette_preparing_data.html) section, include examples of how to run the functions described in sections 3-5 to help the user to understand how to find/specify de DATASETCODE

Great suggestions! Examples are now added.

> In the section [User cases FAOSTAT raw data](https://jgcri.github.io/gcamfaostat/articles/vignette_use_cases.html#faostat-raw-data-and-processing-output-updates), it would be helpful to detail how to change the year variables, since there are many variables that are year related. Maybe include an example? Also, for users not familiarised with the gcamdata pkg, it might be helpful to specify the path where constants.R is located

`constants.R` is updated to `R/constants.R`. This sentence is added: "If more recent data is available the maximum year in `FAOSTAT_Hist_Year_*` can be updated accordingly." We indeed recently updated the data to 2021 (in the base year update branch: GCAM_BYU2021), which will be needed for the planned GCAM updates. We will included additional details when that branch is merged. Thanks

> In the [User cases Country aggregation](https://jgcri.github.io/gcamfaostat/articles/vignette_use_cases.html#country-aggregation-and-disaggregation), would be helpful to detail how to obtain the area_code in the FAOSTAT_AREA_RM_NONEXIST function.

Thanks! area_code was defined by FAO and included in all FAOSTAT datasets. The following sentence is added: "Note the functions rely on the `area_code` ID included in FAO dataset. They are also consistently defined in the function. "
Please also note that in the next version, we will utilize the area_code more broadly. This is already done in a base year update (BYU2021) branch which will be merged later. See issue addressed here https://github.com/JGCRI/gcamfaostat/issues/4

> In the [bilateral trade case study](https://jgcri.github.io/gcamfaostat/articles/vignette_use_cases.html#case-study-wheat-bilateral-trade-pattern-changes), when considering the download of the latest TM, you could include the command

Thanks. The following code was included to download the latest data.
```
FF_download_FAOSTAT(DATASETCODE = "TM",
                    DATA_FOLDER = file.path("..",DIR_RAW_DATA_FAOSTAT),
                    OverWrite = TRUE)
```
But in the BYU branch, we will try to update the prebuilt data it the processing here can be made easier.

> In the [bilateral trade case study](https://jgcri.github.io/gcamfaostat/articles/vignette_use_cases.html#case-study-wheat-bilateral-trade-pattern-changes) I found this message `cannot open destfile '../inst/extdata/aglu/FAO/FAOSTAT/Trade_DetailedTradeMatrix_E_All_Data_(Normalized).zip', reason 'No such file or directory'`
and I made sure that in ../inst/extdata/aglu/FAO/FAOSTAT/Trade_DetailedTradeMatrix_E_All_Data_(Normalized).zip I have the file. The only element that might be different, is that it has the WinRAR ZIP archive (.zip) extension

In the `DATA_FOLDER = file.path("..",DIR_RAW_DATA_FAOSTAT)` as the input to the function, you want to remove `".."` to use `DATA_FOLDER = file.path(DIR_RAW_DATA_FAOSTAT)` unless running in the `vignette` folder. I think it should work.
