### 2024-03-15 realxinzhao/gcamfaostat#11 @realxinzhao
> Check and fix the length of the paper.

@HenriKajasilta thank you so much for pointing this out and for the suggestions. We have now made a significant cut of the manuscript. Visualization and other functions were cut. The Design and Functionality section overall is reduced. However, it is still around 1500 (not including captions/reference). Please let us know if this is fine (given that it's the first of its kind package so we wanted to include more communication to the audience). We are also happy to make more cuts if needed.

> Indicate if there are other packages which are similar to gcamfaostat.

Thanks! We added this sentence to the Statement of Need: "To the best of our knowledge, such a tool has not been developed yet."
