### 2024-03-15 realxinzhao/gcamfaostat#9 @realxinzhao
> It seems that the function (FF_download_RemoteArchive) does not accept a vector as DATASETCODE. Either fix it or document accordingly

Good catch. We fixed the function by adding an assertion to only allow single dataset code input and give a message: "Single dataset allowed; consider using a loop or the function `FF_rawdata_info()` for downloading multiple datasets". The function `FF_download_RemoteArchive` is also corrected accordingly. The corresponding documentation is also updated.

>Consider running the package through the [styler package](https://www.tidyverse.org/blog/2017/12/styler-1.0.0/), with a style of your choice, to ensure clear and consistent code style (e.g. styler:::style_active_pkg())

Great suggestion. We tested this as suggested and the style package works great. However, we realize that there is an active branch, [GCAM_BYU2021](https://github.com/JGCRI/gcamfaostat/tree/GCAM_BYU2021), that will be merged soon (for gcamfaostat v1.1). We will make this full on style updates after that merge to avoid potential conflicts. Thanks!

![Styler](https://github.com/JGCRI/gcamfaostat/assets/47983827/d8fc6acd-e1a8-401f-891b-8a554803ed1c)
