### 2024-03-16 realxinzhao/gcamfaostat#10 @realxinzhao
Thank you for the valuable comments.

It seems there was some confusion caused by the `driver()` function, which was not a preferred approach compared to `driver_drake()`.  This is now clarified in [Getting Started](https://jgcri.github.io/gcamfaostat/articles/vignette_getting_started.html#driver_drake-and-driver). And only `driver_drake` was included in the paper and [Quick Start](https://jgcri.github.io/gcamfaostat/index.html#id_4-run-the-driver).
But in short, both functions were inherited from `gcamdata` and they were documented. For `gcamfaostat`, it is recommended to use [driver_drake()](https://jgcri.github.io/gcamfaostat/reference/driver_drake.html) so the data tracing and exploring after the drive will be possible, even though it may take longer to run and additional space (for drake cache) compared to [driver()](https://jgcri.github.io/gcamfaostat/reference/driver.html).

> Fix the aforementioned workflow.

Please note that `driver_drake()` needs to be run successfully to build a `drake cache`. And the `load_from_cache()` function loads the data from the `drake cache`. It appears the error you shared was `cannot find drake cache`, which is likely because `driver_drake()` was not run successfully. Please check this can let us know if the issue remains.
The issue in #8 was a little different I believe.

> Suggestion: Create an example of how to use the function(s), that is/are easy to perform and light to run.

As explained, the `driver()` function, while working, won't be the main approach for `gcamfaostat`. Many of the options in the `driver()` function can be replaced by the use of the `drake cache` generated by running `driver_drake()`. Thus, the examples we provided focused on utilizing `drake cache` to trace intermediate data flows (e.g., ``load_from_cache()`).

> Fix data storing to multiple locations.

Good catch! Writing out intermediate data to `csv` files was a default option in `driver()`. The following sentence is added to [Getting Started](https://jgcri.github.io/gcamfaostat/articles/vignette_getting_started.html#driver_drake-and-driver):
"The `driver()` approach has a default option (`write_outputs =  T`) to write the intermediate `csv` outputs to `./outputs`"

> Fix the issue so that building the package won't produce an error.

Package checking errors have been addressed. Thank you!
