### 2024-03-16 a89a9d2
Address Code Suggestions #9 and Usability Fixes #10
Clear errors and warning in package checking
