+++ R/constants.R
@@ -61,7 +61,7 @@ Hist_MEAN_Year_NUTRIENT_MASS_CONV <- 2010:2019 # average cal per g
 # decimal places in ggplot
 scaleFUN <- function(x) sprintf("%.0f", x)
 
-
+xml.XML_SUFFIX <- NULL
 
 
 
