+++ DESCRIPTION
@@ -29,7 +29,8 @@ Imports:
     ggplot2 (>= 3.3.2),
     xml2 (>= 1.3.2),
     XML (>= 3.99-0.5),
-    rlang
+    rlang,
+    kableExtra
 Suggests:
     igraph (>= 1.0.1),
     testthat (>= 3.0.0),
@@ -41,6 +42,6 @@ Suggests:
     rmarkdown
 Encoding: UTF-8
 LazyData: true
-RoxygenNote: 7.2.3
+RoxygenNote: 7.3.1
 VignetteBuilder: knitr
 Config/testthat/edition: 3
