### 2024-03-16 realxinzhao/gcamfaostat#9 @realxinzhao
>  Some warnings and notes when performing devtools::check() could be easily removed

Thanks. We have now cleared all the errors. Warnings are also cleared as much as possible.
There are two remaining but unimportant warnings. We double checked out LICENSE specifications, which should be fine. The `qpdf` warning seems to be R version related and being discussed [here](https://github.com/ropensci/qpdf/issues/17).

```
> checking DESCRIPTION meta-information ... WARNING
  Invalid license file pointers: LICENSE

> checking R/sysdata.rda ... OK
   WARNING
  'qpdf' is needed for checks on size reduction of PDFs
```
