### 2024-03-16 #6388 @realxinzhao
Dear @klau506 and @hugoledoux,

Thank you for the careful reading of our paper and package. All the comments are very clear, spot-on, and useful! We believe the paper & package has been significantly improved as a result of the very useful comments and suggestions.

The detailed point-by-point point responses are provided below in the issues suggested below:

[Manuscript Suggestions #7](https://github.com/JGCRI/gcamfaostat/issues/7)
[Documentation Suggestions #8](https://github.com/JGCRI/gcamfaostat/issues/8)
[Code Suggestions #9](https://github.com/JGCRI/gcamfaostat/issues/9)
[Usability Fixes #10](https://github.com/JGCRI/gcamfaostat/issues/10)
[Comments on the paper #11](https://github.com/JGCRI/gcamfaostat/issues/11)

Please let us know if you have more comments and suggestions or if we can provide additional information.

Sincerely,
Xin Zhao (for the authors)
