### 2024-04-03 realxinzhao/gcamfaostat#11 @HenriKajasilta
I would still consider shortening the paper as it is considerably longer than the 1000 word limit. In addition, _Key functions_ sections contain detailed commands on how to run the application, whereas the paper only expects a high-level description, and I would see fit to reduce the description of details here.
