### 2024-04-03 realxinzhao/gcamfaostat#10 @HenriKajasilta
I installed the package from scratch following the instructions in the documentation. I am not very familiar with the _renv_ package beforehand, so this could also be a user error. However, when I follow the instructions for using renv [here](https://jgcri.github.io/gcamdata/articles/getting-started/getting-started.html#optional-renv-currently-included-in-gcam-6-0-and-gcamdata-release) I am not able to get the packages needed to run the application. Can you please check if following these steps gives you the expected outcome? I can make a new issue about this, but first I would like to clarify how it should work.
