### 2024-04-03 realxinzhao/gcamfaostat#10 @realxinzhao
@HenriKajasilta Thank you!
Did you see a specific error message when running `driver_drake`?

I believe `renv` is only to ensure the version of the dependency packages is identical to the ones specified in `renv` files. It is not really needed unless there are package version conflicts that could affect the running of the package.
We tried different machines before usually without needing `renv`.
