### 2024-04-03 realxinzhao/gcamfaostat#10 @HenriKajasilta
Yes @realxinzhao , I have the same error as mentioned earlier on issue [8](https://github.com/JGCRI/gcamfaostat/issues/8), that is why I wanted to proceed with `renv` to overcome this.

```
ℹ Loading gcamfaostat
> driver_drake()
Loading required namespace: drake
GCAM Data System v1.0.0
Found 19 chunks
Found 77 chunk data requirements
Found 43 chunk data products
37 chunk data input(s) not accounted for
Error in `mutate()`:
ℹ In argument: `command = paste0(PACKAGE_NAME, "::PREBUILT_DATA[['", target, "']]")`.
Caused by error:
! `command` must be size 0 or 1, not 99.
```
