+++ R/driver.R
@@ -161,7 +161,7 @@ driver <- function(all_data = empty_data(),
                    return_data_names = union(inputs_of(return_inputs_of),
                                              outputs_of(return_outputs_of)),
                    return_data_map_only = FALSE,
-                   write_outputs = !return_data_map_only,
+                   write_outputs = FALSE,
                    write_xml = write_outputs,
                    outdir = OUTPUTS_DIR, xmldir = XML_DIR,
                    quiet = FALSE,
@@ -474,7 +474,7 @@ driver_drake <- function(
                             outputs_of(return_outputs_of)),
   return_data_map_only = FALSE,
   return_plan_only = FALSE,
-  write_xml = !return_data_map_only,
+  write_xml = FALSE,
   xmldir = XML_DIR,
   quiet = FALSE,
   user_modifications = NULL,
