+++ DESCRIPTION
@@ -20,12 +20,12 @@ Imports:
     assertthat (>= 0.2),
     dplyr (>= 1.0.4),
     methods,
-    magrittr (>= 1.5),
-    tibble (>= 1.1),
-    tidyr (>= 0.7.1),
-    readr (>= 1.3.1),
-    data.table (>= 1.10.4),
-    purrr (>= 0.3.4),
+    magrittr (>= 2.0),
+    tibble (>= 3.1),
+    tidyr (>= 1.1.3),
+    readr (>= 1.4.0),
+    data.table (>= 1.14.0),
+    purrr (>= 1.0.1),
     ggplot2 (>= 3.3.2),
     xml2 (>= 1.3.2),
     XML (>= 3.99-0.5),
@@ -35,7 +35,7 @@ Suggests:
     igraph (>= 1.0.1),
     testthat (>= 3.0.0),
     usethis (>= 1.4.0),
-    drake (>= 6.2.1),
+    drake (>= 7.13.3),
     stringr,
     tidyselect,
     knitr,
