### 2024-04-03 realxinzhao/gcamfaostat#10 @realxinzhao
@HenriKajasilta  Thank you, and apologies for the inconvenience.
I did try a clean install again and didn't see the error. However, I realize `devtools::install_deps()` only updates local package per the specifications in the `DESCRIPTION` file. And we only required lower versions of the packages mostly following the original settings in `gcamdata`.
I have now updated the package version there so that  `devtools::install_deps()`  (or Rstudio) will check those versions.

I think the issue came from either `tibble` or `dplyr` package versions. But please let us know if the error remains. Thank you so much!
