### 2024-04-04 d778ba1
Shorten the paper
