+++ vignettes/references.bib
@@ -386,3 +386,14 @@ publisher: Nature Publishing Group},
 	pages = {160018},
 	file = {Full Text PDF:C\:\\Users\\zhao752\\Zotero\\storage\\AEIVY5BT\\Wilkinson et al. - 2016 - The FAIR Guiding Principles for scientific data ma.pdf:application/pdf},
 }
+@article{zhao2024nc,
+  title={Trade-offs in land-based carbon removal measures under 1.5° C and 2° C futures},
+  author={Zhao, Xin and Mignone, Bryan K and Wise, Marshall A and McJeon, Haewon C},
+  journal={Nature Communications},
+  volume={15},
+  number={1},
+  pages={2297},
+  year={2024},
+  doi = {	doi = {10.1038/sdata.2016.18},},
+  publisher={Nature Publishing Group UK London}
+}
