### 2024-04-04 #6388 @realxinzhao
Dear @hugoledoux, @klau506 and @hugoledoux,

Following suggestions from @hugoledoux, the paper length is greatly reduced. It seems the 1473 word count by the bot was a count in the md file and it included captions, acknowledgment, and url links. Per our check, the word count in the pdf is about 1200 words and even shorter if not accounting for in-text citations. Will this be okay? Thank you!

Best,
Xin
