### 2024-04-04 realxinzhao/gcamfaostat#11 @realxinzhao
@HenriKajasilta
We further simplified the writing and restructured the Design and Functionality section to reduce the length to <1200 if only accounting for the main components. The structure is much more balanced now.  Please take a look when you have a chance. Thanks!
