### 2024-04-05 realxinzhao/gcamfaostat#10 @HenriKajasilta
@realxinzhao Thank you for taking a look.

It seems that the problem still remains for me. I started with the `renv` approach and ran `renv::init( bare = TRUE)`, which I assume worked as intended. Then running `renv::restore()` doesn’t really do anything and informs that the library is already synchronized with the lockfile. I must be missing something simple here.

Nevertheless, if I continue with `devtools::install_deps(dependencies = TRUE)`, I get the to install the listed _imports_ and _suggests_ packages from the `DESCRIPTION` file (and in my case, those went to the Project Library due to the `renv`), but now I again pump in to the error message, which was mentioned above, when I run the `driver_drake()`.
