### 2024-04-05 realxinzhao/gcamfaostat#10 @realxinzhao
Thank you so much! @HenriKajasilta
I'm testing this on another machine. I did see requests of updating packages
![image](https://github.com/JGCRI/gcamfaostat/assets/47983827/a8e804d0-201a-40ad-8184-0d4901a08daf)

And after updating packages, I did see the error... which indeed because your packages (likely `dplyr` or `tibble`) were newer not older. Sorry about that (I thought I used the most recent versions).
![image](https://github.com/JGCRI/gcamfaostat/assets/47983827/021cd9f8-bbf9-4aa8-b0e3-bab971c5d9b6)

I'm running a `driver()` test, which seems to be fine. I will debug the `mutate` error in `driver_drake` in a bit. Thank you!
