### 2024-04-05 d963d10
Update driver_drake to check if NULL in PREBUILT_DATA
