### 2024-04-05 realxinzhao/gcamfaostat#10 @realxinzhao
Dear @HenriKajasilta
Thanks again for careful checking the package.

We have now fixed the bug which seems to be environment related (`tibble` won't call data in parent `env` if it is not exist in the current package `env`). The small fixe is pushed up. In addition, we also re-initialized the `renv` to update the lock file. I think `renv` won't be needed in your case (since you have the newer packages).
Please let us know if the issues remain or there are other issues. Thank you! And apologies about the inconvenience.
