+++ vignettes/references.bib
@@ -2,6 +2,7 @@
   author       = {Xin Zhao},
   title        = {FAOSTAT AgLU data Archive GCAMv7},
   url={https://doi.org/10.5281/zenodo.8260225},
+  doi={10.5281/zenodo.8260225},
   month        = aug,
   year         = 2023,
   publisher    = {Zenodo},
@@ -58,7 +59,7 @@ publisher: Nature Publishing Group},
 	number=6,
 	language={en},
 	note={number: 6
-publisher: Nature Publishing Group},
+	publisher: Nature Publishing Group},
 	title={A multimodel analysis of post-Glasgow climate targets and feasibility challenges},
 	url={https://www.nature.com/articles/s41558-023-01661-0},
 	volume=13,
@@ -309,6 +310,7 @@ publisher: Nature Publishing Group},
   title={gcamdata: An R package for preparation, synthesis, and tracking of input data for the GCAM integrated human-earth systems model},
   author={Bond-Lamberty, Ben and Dorheim, Kalyn and Cui, Ryna and Horowitz, Russell and Snyder, Abigail and Calvin, Katherine and Feng, Leyang and Hoesly, Rachel and Horing, Jill and Kyle, G Page and others},
   journal={Journal of Open Research Software},
+  doi={0.5334/jors.232},
   volume={7},
   number={1},
   year={2019},
@@ -334,18 +336,9 @@ publisher: Nature Publishing Group},
   year={2018},
   publisher={Pacific Northwest National Lab.(PNNL), Richland, WA (United States)}
 }
-@article{wilkinson2016fair,
-  title={The {FAIR} Guiding Principles for scientific data management and stewardship},
-  author={Wilkinson, Mark D and Dumontier, Michel and Aalbersberg, IJsbrand Jan and Appleton, Gabrielle and Axton, Myles and Baak, Arie and Blomberg, Niklas and Boiten, Jan-Willem and da Silva Santos, Luiz Bonino and Bourne, Philip E and others},
-  journal={Scientific data},
-  volume={3},
-  number={1},
-  pages={1-9},
-  year={2016},
-  publisher={Nature Publishing Group},
-  DOI={10.1038/sdata.2016.18}
-}
-@article{zhaoLandCDR2024,
+
+
+@article{zhao2024nc,
 	title = {Trade-offs in land-based carbon removal measures under 1.5 °{C} and 2 °{C} futures},
 	volume = {15},
 	copyright = {2024 The Author(s)},
@@ -362,8 +355,7 @@ publisher: Nature Publishing Group},
 	year = {2024},
 	note = {Publisher: Nature Publishing Group},
 	keywords = {Climate-change mitigation, Climate-change policy, Environmental impact, Geography},
-	pages = {2297},
-	file = {Full Text PDF:C\:\\Users\\zhao752\\Zotero\\storage\\I9GGJKZ8\\Zhao et al. - 2024 - Trade-offs in land-based carbon removal measures u.pdf:application/pdf},
+	pages = {2297}
 }
 
 @article{wilkinson_fair_2016,
@@ -383,17 +375,6 @@ publisher: Nature Publishing Group},
 	year = {2016},
 	note = {Publisher: Nature Publishing Group},
 	keywords = {Publication characteristics, Research data},
-	pages = {160018},
-	file = {Full Text PDF:C\:\\Users\\zhao752\\Zotero\\storage\\AEIVY5BT\\Wilkinson et al. - 2016 - The FAIR Guiding Principles for scientific data ma.pdf:application/pdf},
+	pages = {160018}
 }
-@article{zhao2024nc,
-  title={Trade-offs in land-based carbon removal measures under 1.5° C and 2° C futures},
[…]
