### 2024-04-07 #6388 @HenriKajasilta
Hi @realxinzhao and @hugoledoux ,

I have reviewed the package code and documentation once more, and it seems that everything is now working as expected. Thank you for looking at the suggestions and improving the functionality. The authors have done a good job with the package and I don’t have any new suggestions to make.
