+++ vignettes/references.bib
@@ -310,7 +310,7 @@ publisher: Nature Publishing Group},
   title={gcamdata: An R package for preparation, synthesis, and tracking of input data for the GCAM integrated human-earth systems model},
   author={Bond-Lamberty, Ben and Dorheim, Kalyn and Cui, Ryna and Horowitz, Russell and Snyder, Abigail and Calvin, Katherine and Feng, Leyang and Hoesly, Rachel and Horing, Jill and Kyle, G Page and others},
   journal={Journal of Open Research Software},
-  doi={0.5334/jors.232},
+  doi={10.5334/jors.232},
   volume={7},
   number={1},
   year={2019},
