### 2024-04-12 #6388 @realxinzhao
@hugoledoux Thank you so much!

The released version on GitHub is `gcamfaostat v1.0.0`: https://github.com/JGCRI/gcamfaostat/releases/tag/v1.0.0
The Zenodo archive is created with metadata updated: https://zenodo.org/records/10967067
The DOI is [10.5281/zenodo.10967067](https://doi.org/10.5281/zenodo.10967067)

Please let us know if we can provide more information. Thanks!
