### 2024-04-14 e24a6b8
Add a new export data of capital stock for GCAM KLEAM
