+++ R/xfaostat_L999_CSVExport.R
@@ -18,7 +18,8 @@
 module_xfaostat_L999_CSVExport <- function(command, ...) {
 
   MODULE_INPUTS <-
-    c("PD",
+    c("CS",
+      "PD",
       "Bal_new_all",
       "FBSH_CB_wide",
       "QCL_PROD",
@@ -44,6 +45,7 @@ module_xfaostat_L999_CSVExport <- function(command, ...) {
       CSV = "GCAMFAOSTAT_AnimalStock",
       CSV = "GCAMFAOSTAT_ProdPrice",
       CSV = "FAO_GDP_Deflators",
+      CSV = "GCAMFAOSTAT_CapitalStock",
       CSV = "GCAMFAOSTAT_MacroNutrientRate",
       CSV = "GCAMFAOSTAT_ForProdTrade",
       CSV = "GCAMFAOSTAT_ForExpPrice",
@@ -297,7 +299,7 @@ module_xfaostat_L999_CSVExport <- function(command, ...) {
 
       rm(QCL_CROP_PRIMARY, QCL_FODDERCROP, QCL_PROD)
 
-      #**AnimalStocks ----
+      ## *AnimalStocks ----
       # Laying is not needed for now; live animal stocks and milk animals are used for water demand calculation in GCAM
       # 1171 "Animals live nes" and 1083 "Pigeons, other birds" are not available
 
@@ -333,7 +335,7 @@ module_xfaostat_L999_CSVExport <- function(command, ...) {
       )
 
       # Producer prices in quantity and value ----
-      # **Producer Price ----
+      ## *Producer Price ----
 
       QCL_PRIMARY_PROD_PV %>%
         filter(year >= MIN_HIST_PP_YEAR) %>%
@@ -361,7 +363,7 @@ module_xfaostat_L999_CSVExport <- function(command, ...) {
       rm(QCL_PRIMARY_PROD_PV)
 
 
-      ###** GDP deflators ----
+      ## *GDP deflators ----
 
       FAO_GDP_Deflators <-
         PD %>% FAOSTAT_AREA_RM_NONEXIST(RM_AREA_CODE = NULL) %>%
@@ -426,7 +428,7 @@ module_xfaostat_L999_CSVExport <- function(command, ...) {
 
       # Forestry data output ----
 
-      #*ForProdTrade ----
+      ## *ForProdTrade ----
       For_Balance %>%
         filter(element != "Demand") %>%
         spread(year, value) %>%
@@ -450,7 +452,7 @@ module_xfaostat_L999_CSVExport <- function(command, ...) {
         out_dir = DIR_OUTPUT_CSV
       )
 
-      #*ForExpPrice ----
+      ## *ForExpPrice ----
       FO_RoundwoodProducts_Export_Q_V %>%
         spread(year, value) %>%
         add_title("GCAMFAOSTAT_ForExpPrice") %>%
@@ -476,7 +478,7 @@ module_xfaostat_L999_CSVExport <- function(command, ...) {
 
       # Land Cover data output ----
 
-      #*LandCover ----
+      ## *LandCover ----
       RL_LandCover %>%
         spread(year, value) %>%
         add_title("GCAMFAOSTAT_LandCover") %>%
@@ -502,7 +504,7 @@ module_xfaostat_L999_CSVExport <- function(command, ...) {
 
       # Fertilizer data output----
 
-      #* N Fertilizer ----
+      ## *N Fertilizer ----
       RFN_ProdDemand %>%
         spread(year, value) %>%
         add_title("GCAMFAOSTAT_NFertilizer") %>%
@@ -536,6 +538,35 @@ module_xfaostat_L999_CSVExport <- function(command, ...) {
         code = "NA",
[…]
