### 2024-04-19 03623dd
Update names in paper.md to be consistent with zenodo
