### 2024-04-19 #6388 @hugoledoux
no: "ChrisT Vernon" vs "Chris Vernon" 🫤
