### 2024-04-20 #6388 @kthyng
Paper comments:
* page 3: paragraph before last paragraph — looks like some of the markdown is off like is one of these supposed to be a link but there is a space so it isn't coming through correctly? Including the parentheses in the links looks strange too.
* Also please check the capitalization in your references. You can preserve capitalization by placing {} around characters/words in your .bib file.
