+++ vignettes/references.bib
@@ -288,7 +288,7 @@ publisher: Nature Publishing Group},
 }
 
 @article{zhao_cmp360,
-  title={Core Model Proposal\# 360: GCAM agriculture and land use (AgLU) data and method updates: connecting land hectares to food calories},
+  title={Core {M}odel {P}roposal 360: GCAM agriculture and land use (AgLU) data and method updates: connecting land hectares to food calories},
   author={Zhao, Xin and Wise, Marshall},
   number={PNNL-34313},
   year={2023},
@@ -307,7 +307,7 @@ publisher: Nature Publishing Group},
   DOI={10.5194/gmd-12-677-2019}
 }
 @article{bond2019gcamdata,
-  title={gcamdata: An R package for preparation, synthesis, and tracking of input data for the GCAM integrated human-earth systems model},
+  title={{g}camdata: An R package for preparation, synthesis, and tracking of input data for the GCAM integrated human-earth systems model},
   author={Bond-Lamberty, Ben and Dorheim, Kalyn and Cui, Ryna and Horowitz, Russell and Snyder, Abigail and Calvin, Katherine and Feng, Leyang and Hoesly, Rachel and Horing, Jill and Kyle, G Page and others},
   journal={Journal of Open Research Software},
   doi={10.5334/jors.232},
