### 2024-04-20 #6388 @realxinzhao
Thank you so much, @kthyng!!

- The formatting in that paragraph is now fixed.
- Two places in `.bib` file is also fixed (with better capitalizations).
