+++ vignettes/references.bib
@@ -201,7 +201,7 @@ publisher: Nature Publishing Group},
 	number=1,
 	language={en},
 	note={number: 1},
-	title={Incorporating Nutritional Accounts to the GTAP Data Base},
+	title={Incorporating Nutritional Accounts to the {GTAP} Data Base},
 	url={https://jgea.org/ojs/index.php/jgea/article/view/150},
 	volume=7,
 	author={Chepeliev, Maksym},
@@ -228,7 +228,7 @@ publisher: Nature Publishing Group},
 	number=1,
 	language={en},
 	note={_eprint: https://onlinelibrary.wiley.com/doi/pdf/10.1111/agec.12086},
-	title={Why do global long-term scenarios for agriculture differ? An overview of the AgMIP Global Economic Model Intercomparison},
+	title={Why do global long-term scenarios for agriculture differ? An overview of the {AgMIP} Global Economic Model Intercomparison},
 	url={https://onlinelibrary.wiley.com/doi/abs/10.1111/agec.12086},
 	volume=45,
 	author={von Lampe, Martin and Willenbockel, Dirk and Ahammad, Helal and Blanc, Elodie and Cai, Yongxia and Calvin, Katherine and Fujimori, Shinichiro and Hasegawa, Tomoko and Havlik, Petr and Heyhoe, Edwina and Kyle, Page and Lotze-Campen, Hermann and Mason d'Croz, Daniel and Nelson, Gerald C. and Sands, Ronald D. and Schmitz, Christoph and Tabeau, Andrzej and Valin, Hugo and van der Mensbrugghe, Dominique and van Meijl, Hans},
@@ -245,7 +245,7 @@ publisher: Nature Publishing Group},
 	year=2023,
 }
 
-@article{Landau2018, doi = {10.21105/joss.00550}, url = {https://doi.org/10.21105/joss.00550}, year = {2018}, publisher = {The Open Journal}, volume = {3}, number = {21}, pages = {550}, author = {William Michael Landau}, title = {The drake R package: a pipeline toolkit for reproducibility and high-performance computing}, journal = {Journal of Open Source Software} }
+@article{Landau2018, doi = {10.21105/joss.00550}, url = {https://doi.org/10.21105/joss.00550}, year = {2018}, publisher = {The Open Journal}, volume = {3}, number = {21}, pages = {550}, author = {William Michael Landau}, title = {The drake {R} package: a pipeline toolkit for reproducibility and high-performance computing}, journal = {Journal of Open Source Software} }
 
 @software{bond_lamberty_2023,
   author       = {Ben Bond-Lamberty and
@@ -278,7 +278,7 @@ publisher: Nature Publishing Group},
                   Marshall Wise and
                   Xin Zhao and
                   Maridee Weber},
-  title        = {JGCRI/gcam-core: GCAM 7.0},
+  title        = {JGCRI/gcam-core: {GCAM} 7.0},
   month        = jun,
   year         = 2023,
   publisher    = {Zenodo},
@@ -307,7 +307,7 @@ publisher: Nature Publishing Group},
   DOI={10.5194/gmd-12-677-2019}
 }
 @article{bond2019gcamdata,
-  title={{g}camdata: An R package for preparation, synthesis, and tracking of input data for the GCAM integrated human-earth systems model},
+  title={{g}camdata: An {R} package for preparation, synthesis, and tracking of input data for the GCAM integrated human-earth systems model},
   author={Bond-Lamberty, Ben and Dorheim, Kalyn and Cui, Ryna and Horowitz, Russell and Snyder, Abigail and Calvin, Katherine and Feng, Leyang and Hoesly, Rachel and Horing, Jill and Kyle, G Page and others},
   journal={Journal of Open Research Software},
   doi={10.5334/jors.232},
