### 2024-04-22 #6388 @kthyng
Ok good @realxinzhao! That R is capitalized, though I see a number of other proper nouns in references that should still be capitalized.
