+++ vignettes/references.bib
@@ -60,7 +60,7 @@ publisher: Nature Publishing Group},
 	language={en},
 	note={number: 6
 	publisher: Nature Publishing Group},
-	title={A multimodel analysis of post-Glasgow climate targets and feasibility challenges},
+	title={A multimodel analysis of post-{G}lasgow climate targets and feasibility challenges},
 	url={https://www.nature.com/articles/s41558-023-01661-0},
 	volume=13,
 	author={van de Ven, Dirk-Jan and Mittal, Shivika and Gambhir, Ajay and Lamboll, Robin D. and Doukas, Haris and Giarola, Sara and Hawkes, Adam and Koasidis, Konstantinos and Köberle, Alexandre C. and McJeon, Haewon and Perdana, Sigit and Peters, Glen P. and Rogelj, Joeri and Sognnaes, Ida and Vielle, Marc and Nikas, Alexandros},
@@ -93,7 +93,7 @@ publisher: Nature Publishing Group},
 	number=4,
 	language={en},
 	note={_eprint: https://onlinelibrary.wiley.com/doi/pdf/10.1029/2022EF003063},
-	title={Trade and Climate Mitigation Interactions Create Agro-Economic Opportunities With Social and Environmental Trade-Offs in Latin America and the Caribbean},
+	title={{T}rade and {C}limate {M}itigation {I}nteractions {C}reate {Agro-Economic} {O}pportunities {W}ith {S}ocial and {E}nvironmental {Trade-Offs} in {L}atin {A}merica and the {C}aribbean},
 	url={https://onlinelibrary.wiley.com/doi/abs/10.1029/2022EF003063},
 	volume=11,
 	author={Yarlagadda, Brinda and Wild, Thomas and Zhao, Xin and Clarke, Leon and Cui, Ryna and Khan, Zarrar and Birnbaum, Abigail and Lamontagne, Jonathan},
@@ -167,7 +167,7 @@ publisher: Nature Publishing Group},
 	issn={0013-936X},
 	number=19,
 	note={publisher: American Chemical Society},
-	title={FABIO—The Construction of the Food and Agriculture Biomass Input–Output Model},
+	title={FABIO—{T}he {C}onstruction of the {F}ood and {A}griculture {B}iomass {I}nput–{O}utput {M}odel},
 	url={https://doi.org/10.1021/acs.est.9b03554},
 	volume=53,
 	author={Bruckner, Martin and Wood, Richard and Moran, Daniel and Kuschnig, Nikolas and Wieland, Hanspeter and Maus, Victor and Börner, Jan},
@@ -201,7 +201,7 @@ publisher: Nature Publishing Group},
 	number=1,
 	language={en},
 	note={number: 1},
-	title={Incorporating Nutritional Accounts to the {GTAP} Data Base},
+	title={{I}ncorporating {N}utritional {A}ccounts to the {GTAP} {D}ata {B}ase},
 	url={https://jgea.org/ojs/index.php/jgea/article/view/150},
 	volume=7,
 	author={Chepeliev, Maksym},
@@ -212,7 +212,7 @@ publisher: Nature Publishing Group},
 	day=25,
 }
 @article{zhao2021role,
-  title={The role of global agricultural market integration in multiregional economic modeling: Using hindcast experiments to validate an Armington model},
+  title={The role of global agricultural market integration in multiregional economic modeling: Using hindcast experiments to validate an {Armington} model},
   author={Zhao, Xin and Calvin, Katherine V and Wise, Marshall A and Iyer, Gokul},
   journal={Economic Analysis and Policy},
[…]
