### 2024-04-22 #6388 @realxinzhao
Oops... yes! @kthyng
I think they are fixed now (the titles all match the ones that appear in their journal articles now). Thank you!
