### 2024-04-22 #6388 @realxinzhao
> Congrats on your new publication @realxinzhao! Many thanks to editor @hugoledoux and reviewers @klau506 and @HenriKajasilta for your time, hard work, and expertise!!

Thank you all so much, All! This experience of leading a software publication as an economist/human-earth system modeler has been incredibly rewarding. I've learned a tremendous amount from the process, and I deeply appreciate the time and effort you've dedicated to enhancing both the software and the paper.

Best regards,
Xin
